Improve names of extension files
