+++ base/src/main/kotlin/io/spine/protobuf/AnyExts.kt
