+++ base/src/main/kotlin/io/spine/protobuf/DescriptorExts.kt
